Extracts main blocks of tag controller and spec code to concerns. Adds to orgs for indexing/searching.
