+++ spec/controllers/api/v1/organizations_controller_spec.rb
@@ -2,6 +2,7 @@ require 'spec_helper'
 
 describe Api::V1::OrganizationsController, type: :controller do
   let(:authorized_user) { create(:user) }
+  let(:api_resource_name) { "organizations" }
   let(:organization) { build(:organization, listed_at: Time.now, owner: authorized_user) }
   let(:unlisted_organization) { build(:unlisted_organization) }
   let(:owned_unlisted_organization) { build(:unlisted_organization, owner: authorized_user) }
@@ -63,6 +64,11 @@ describe Api::V1::OrganizationsController, type: :controller do
           get :index
           expect(json_response["organizations"]).to be_empty
         end
+      end
+
+      it_behaves_like "taggable" do
+        let(:resource) { organization }
+        let(:second_resource) { build(:organization, listed_at: Time.now, owner: authorized_user) }
       end
 
       describe "filtering by slug" do
